+++ src/ca/qc/bdeb/vue/Affichable.java
@@ -11,7 +11,7 @@ package ca.qc.bdeb.vue;
 public interface Affichable {
     
     public String getNomImg();
-    public int getX();
-    public int getY();
+    public double getX();
+    public double getY();
     
 }
